Test cases issues resolved

+++ app/src/main/java/com/deejayen/note/database/entity/Note.kt
@@ -7,6 +7,6 @@ import androidx.room.PrimaryKey
 
 @Entity
 data class Note(
-    @PrimaryKey val noteId: Int? = null,
+    @PrimaryKey val noteId: Long? = null,
     val title: String,
 )
